[Snippets] Refactored Loop related classes
