+++ src/common/snippets/include/snippets/lowered/pass/split_loops.hpp
@@ -36,8 +36,7 @@ public:
     bool run(LinearIR& linear_ir, lowered::LinearIR::constExprIt begin, lowered::LinearIR::constExprIt end) override;
 
 private:
-    static bool can_be_split(const LinearIR::LoopManager::LoopInfoPtr& current,
-                             const LinearIR::LoopManager::LoopInfoPtr& target);
+    static bool can_be_split(const LoopInfoPtr& current, const LoopInfoPtr& target);
 };
 
 } // namespace pass
