+++ src/common/snippets/src/lowered/pass/iter_handler.cpp
@@ -91,7 +91,7 @@ bool TransformInnerSplitLoop::run(LinearIR& linear_ir, LinearIR::constExprIt beg
     const auto& loop_manager = linear_ir.get_loop_manager();
     const auto& loop_info = loop_manager->get_loop_info(loop_end->get_id());
     const auto current_dim_idx = loop_info->get_dim_idx();
-    OPENVINO_ASSERT(current_dim_idx != LinearIR::LoopManager::LoopInfo::UNDEFINED_DIM_IDX,
+    OPENVINO_ASSERT(current_dim_idx != LoopInfo::UNDEFINED_DIM_IDX,
                     "Outer splitted loop unexpectedly iterates by several dimension indices");
 
     bool modified = false;
